requirements.txt
pandas==2.2.3
pyspark==3.5.4
delta-spark==3.3.2